Split Order Provider into Controller/Provider
